# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MTC App Api",
        "description": "MTC Application OpenApi Documentation",
        "contact": {
            "email": "aminetemadifar.w@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "localhost"
        }
    ],
    "paths": {
...
        "/ranks": {
            "get": {
                "tags": [
                    "Ranks"
                ],
                "summary": "Get all ranks",
                "description": "Retrieve a list of all ranks",
                "operationId": "53431c9564c31082957233c65930e7d3",
                "responses": {
                    "200": {
                        "description": "List of ranks retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RankResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ranks"
                ],
                "summary": "Create a new rank",
                "description": "Create a new rank with the provided data",
                "operationId": "92398227183e27f218c07f5a86f65b95",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Instructor"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rank created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RankResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/ranks/{rank}": {
            "get": {
                "tags": [
                    "Ranks"
                ],
                "summary": "Get a specific rank",
                "description": "Retrieve details of a specific rank",
                "operationId": "cd739995932f7407e99af8cd5567d655",
                "parameters": [
                    {
                        "name": "rank",
                        "in": "path",
                        "description": "Rank ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rank details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RankResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rank not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rank not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Ranks"
                ],
                "summary": "Update a rank",
                "description": "Update the details of a specific rank",
                "operationId": "2195a5a65bcc82efde4becb4e69c1975",
                "parameters": [
                    {
                        "name": "rank",
                        "in": "path",
                        "description": "Rank ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Senior Instructor"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rank updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RankResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rank not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rank not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Ranks"
                ],
                "summary": "Delete a rank",
                "description": "Delete a specific rank",
                "operationId": "56ea4245961515c0bf0b823996922cd5",
                "parameters": [
                    {
                        "name": "rank",
                        "in": "path",
                        "description": "Rank ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rank deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rank deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rank not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rank not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "RankResource": {
                "title": "Rank Resource",
                "description": "Rank resource schema",
                "properties": {
                    "id": {
                        "description": "Rank code",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Rank name",
                        "type": "string",
                        "example": "Beginner"
                    },
                    "status": {
                        "description": "Rank status",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}